requirements.txt
torch
pylzma
tokenizers
tqdm
transformers
flask
flask-cors
python-dotenv
pickle
requests